requirements.txt
django>=4.2.1,<5.0.0
djangorestframework>=3.14.0,<4.0.0
